requirement.txt
numpy
opencv-python
pillow
pdf2image
ultralytics
pytesseract
concurrent.futures
multiprocessing
tqdm
argparse

sudo apt install poppler-utils
sudo apt install tesseract-ocr